start working on HunchBack

+++ C64Lib/Sprites.cs
@@ -20,7 +20,7 @@ namespace C64Lib
         public bool IsInCollision { get; set; }
         public Colors Color { get; set; }
         public bool MultiColor { get; set; }
-        public int DataBlock { get; set; }
+        public ulong DataBlock { get; set; }
         public uint X { get; set; }
         public uint Y { get; set; }
